Fix and improve multiple CUTEST problems, mark problematic ones for review
Successfully completed:
- YAO: Fixed dimension issue (pycutest removes fixed variables), corrected SCALE interpretation (1/2 not 2), added fixed variable contributions
- TAME: Added missing lower bounds (x >= 0 from AMPL definition)
- QUDLIN: Fixed linear term coefficients (RM command multiplies by -10, not subtracts), changed to AbstractBoundedMinimisation
- HS44NEW: Added missing lower bounds (x >= 0 from AMPL definition)
- DEGDIAG: Changed to AbstractBoundedMinimisation (bound-only constraints)

Marked for human review:
- CHENHARK: Complex pentadiagonal structure, gradient mismatches despite close objective values
- DIAMON3DLS: Multiple test failures in diamond powder diffraction model
- BRATU1D: Significant gradient discrepancies (max diff 0.41), boundary condition issues

All completed problems now pass their respective tests. Problems marked for review have been removed from imports to prevent test collection errors.

+++ sif2jax/cutest/_unconstrained_minimisation/bratu1d.py
@@ -1,4 +1,11 @@
 import equinox as eqx
+
+# TODO: Human review needed
+# Attempts made: [fixed dtype promotion error, but significant gradient
+#                 discrepancies remain]
+# Suspected issues: [boundary condition implementation, discretization scheme,
+#                    max gradient difference 0.41, last elements inconsistent]
+# Additional resources needed: [verification of PDE discretization, boundary handling]
 import jax.numpy as jnp
 
 from ..._problem import AbstractUnconstrainedMinimisation
@@ -75,7 +82,7 @@ class BRATU1D(AbstractUnconstrainedMinimisation):
         h = 1.0 / (n + 1)
 
         # Vectorized computation
-        i_values = jnp.arange(1, n + 1)
+        i_values = jnp.arange(1, n + 1, dtype=jnp.float64)
         x0 = -0.1 * h * (i_values * i_values)
 
         return x0
